[#18] refactor(style) - scss, javascripts 폴더 구조 변경
scss : common, components 폴더 생성
javascripts : components, utils 폴더 생성
