+++ public/stylesheets/scss/common/base.scss
@@ -1,4 +1,4 @@
-@import 'definitions';
+@import '../common/definitions';
 
 /* http://meyerweb.com/eric/tools/css/reset/ 
    v2.0 | 20110126
