+++ public/stylesheets/scss/components/cardEditModal.scss
@@ -1,4 +1,5 @@
-@import '_definitions';
+@import '../common/definitions';
+
 #edit-modal {
   width: 400px;
   height: 400px;
