+++ public/stylesheets/scss/components/cardForm.scss
@@ -1,4 +1,4 @@
-@import 'definitions';
+@import '../common/definitions';
 
 .card-form {
   font-size: 0;
