+++ public/stylesheets/scss/components/column.scss
@@ -1,4 +1,4 @@
-@import 'definitions';
+@import '../common/definitions';
 
 #column-container {
   display: flex;
