+++ public/stylesheets/scss/components/header.scss
@@ -1,4 +1,5 @@
-@import '_definitions';
+@import '../common/definitions';
+
 #head {
   display: flex;
   align-items: center;
