+++ public/stylesheets/scss/components/sidebar.scss
@@ -1,4 +1,5 @@
-@import 'definitions';
+@import '../common/definitions';
+
 #sidebar {
   height: calc(100vh - 60px);
   display: flex;
