+++ views/CardEditModal.pug
@@ -1,5 +1,5 @@
 head
-    link(rel='stylesheet' href='/static/stylesheets/css/cardEditModal.css')
+    link(rel='stylesheet' href='/static/stylesheets/css/components/cardEditModal.css')
 body
     #edit-modal
         .modal-header
