+++ views/Sidebar.pug
@@ -6,4 +6,4 @@
         img(src='/static/icon/notification.png')
         span Activity
     include ./SidebarCard.pug
-script(src="/static/javascripts/SideBar.js",type="module")
+script(src="/static/javascripts/components/SideBar.js",type="module")
