+++ views/index.pug
@@ -2,12 +2,12 @@ doctype html
 html
     head
         meta(charset='UTF-8')
-        link(rel='stylesheet' href='/static/stylesheets/css/base.css')
-        link(rel='stylesheet' href='/static/stylesheets/css/header.css')
-        link(rel='stylesheet' href='/static/stylesheets/css/sidebar.css')
-        link(rel='stylesheet' href='/static/stylesheets/css/cardEditModal.css')
-        link(rel='stylesheet' href='/static/stylesheets/css/column.css')
-        link(rel='stylesheet' href='/static/stylesheets/css/cardForm.css')
+        link(rel='stylesheet' href='/static/stylesheets/css/common/base.css')
+        link(rel='stylesheet' href='/static/stylesheets/css/components/header.css')
+        link(rel='stylesheet' href='/static/stylesheets/css/components/sidebar.css')
+        link(rel='stylesheet' href='/static/stylesheets/css/components/cardEditModal.css')
+        link(rel='stylesheet' href='/static/stylesheets/css/components/column.css')
+        link(rel='stylesheet' href='/static/stylesheets/css/components/cardForm.css')
         title 메인페이지
     body
         include ./Header.pug
